SI2 V0.11.2 - Disparos desde los enemigos con colision disparo a jugador, arreglado problema de los disparos al eliminar enemigo. Falta aniadir explosion.

+++ core/src/com/politecnicomalaga/sp2/model/HeroBullet.java
@@ -41,7 +41,7 @@ public class HeroBullet extends Actor {
     }//RESETPOS
 
     public Rectangle getBody() {
-        return body = new Rectangle((int) getX(), (int) getY(), (int) getWidth(), (int) getHeight());
+        return body = new Rectangle(getX(), getY(), getWidth(), getHeight());
     }//GETBODY
 
     @Override
